==> 1 prepare_data.ipynb <==
## Импортирование библиотек

import pandas as pd
import numpy as np
import os
from tqdm.auto import tqdm
from scipy.sparse import csr_matrix
import gzip
import pickle

## Распаковка набора данных

!unzip archive.zip

!ls

## Проверка числа уникальных значений в столбцах

colmaps = {}
for file in tqdm([file for file in os.listdir('./') if file.startswith('part')]):
    df = pd.read_parquet(file)
    for col in tqdm(df.columns):
        if col not in colmaps:
            colmaps[col] = {}
        for val in df[col].unique():
            if val in colmaps[col]: continue
            colmaps[col][val] = len(colmaps[col])

for k, v in colmaps.items():
    print(k, len(v))

## Сохранение индексов уникальных признаков

os.makedirs('files', exists_ok=True)

for k, v in colmaps.items():
    pd.DataFrame({
                  '%s_idx'%k: np.arange(len(v)),
                  k: list(v.keys()),
                 }).to_csv('files/%s_mapper.tsv.gz'%k,
                           sep='\t',
                           index=False,
                           encoding='utf-8',
                           compression='gzip')

## Переконвертация набора данных в индексы и pickle формат

%%time
for file in tqdm([file for file in os.listdir('./') if file.startswith('part')]):
    df = pd.read_parquet(file)
    df2 = df.copy()
    for col in tqdm(df.columns):
        if col == 'user_id': continue
        map_df = pd.read_csv('files/%s_mapper.tsv.gz'%col, sep='\t', dtype=str)
        df2[col] = df2[col].astype(str)
        df2 = df2.merge(map_df, on=col, how='left').drop(col, axis=1)
    df2 = df2.groupby('user_id').agg(list).reset_index()
    df2.to_pickle('files/%s.pickle.gz'%file.split('.')[0])

## Загрузка и объединение сконвертированных файлов в один

%%time
users = None
for file in tqdm([file for file in os.listdir('./') if file.startswith('part')]):
    if users is not None:
        users = pd.concat([users,
                           pd.read_pickle('files/%s.pickle.gz'%file.split('.')[0])]
                         )
    else:
        users = pd.read_pickle('files/%s.pickle.gz'%file.split('.')[0])

## Сброс индекса после конкатенации

users = users.reset_index(drop=True)

## Считывание нужных мапок

# было неправильное понимание порядка дня утро-день-вечер-ночь
# в files/part_of_day_mapper.tsv.gz
# на самом деле верный порядок ночь-утро-день-вечер
# поэтому захардкожено
part_of_day_remap = [1, 2, 3, 0] 
request_cnt_remap = pd.read_csv('files/request_cnt_mapper.tsv.gz', sep='\t').request_cnt.tolist()

## Вывод набора данных

users

## Считываение конвертора дат в правильные индексы 

date_mapper = \
{k:i for i, k in enumerate(
    pd.read_csv('files/date_mapper.tsv.gz', sep='\t',
            dtype=str).sort_values(by=['date'],
                                   ascending=True).date_idx.values.copy()
)}

## Восстановление значений колонки request_cnt

%%time
users['request_cnt'] = \
users.request_cnt_idx.apply(lambda x: [request_cnt_remap[int(xx)] for xx in x])

## Проверка, что число значений в последовательности ссылок и кол-ва запросов совпадает

assert (users['request_cnt'].apply(len) == users['url_host_idx'].apply(len)).all()

## Сохранение мешков слов с учётом request_cnt



def fillna(row):
    vals = np.array(row, dtype='<U')
    mask = vals == 'nan'
    vals[mask] = '-1'
    vals = vals.astype(int)
    pbar.update(1)
    return list(vals)

uids = users.user_id.values.copy()
cnts = users['request_cnt'].values.copy()
for col in tqdm(
            ['region_name_idx', 'city_name_idx',
             'cpe_manufacturer_name_idx', 'cpe_model_name_idx',
             'cpe_type_cd_idx', 'cpe_model_os_type_idx', 'price_idx',
             'date_idx', 'part_of_day_idx']
    ):
    pbar = tqdm(total=users.shape[0])
    vals = users[col].apply(fillna)

    max_val = vals.apply(max).max()+1

    indices = []
    jindices = []
    data = []

    for i, row in enumerate(tqdm(vals)):
        inds = zip([(xx if xx>=0 else max_val) for xx in row], cnts[i])
        inds = sorted(inds, key=lambda x: x[0])
        indices.extend([i]*len(row))
        jindices.extend(list(map(lambda x: x[0], inds)))
        data.extend(list(map(lambda x: x[1], inds)))

    print(len(indices))
    # csr_matrix автоматом просуммирует значение дубликатов по i, j в одно
    XX = \
    csr_matrix((data, (indices, jindices)),
               shape=(users.shape[0], max_val+1),
               dtype=np.uint32)

    with gzip.open('files/%s_cbag_v2.pickle.gz'%(col.split('_idx')[0]), 'wb') as f:
        pickle.dump(dict(data=XX, uids=uids), f, protocol=-1)

## Восстановление упорядоченных индексов для дат и времени суток

%%time
users['date_ordered_idx'] = users.date_idx.apply(lambda x: [date_mapper[xx] for xx in x])
users['part_ordered_idx'] = users.part_of_day_idx.apply(lambda x: [part_of_day_remap[int(xx)] for xx in x])
users

## Получение одной колонки со временем

%%time
users['date_sorting_val'] = users.apply(lambda row: [x*10+y for x,y in 
                                                     zip(row.date_ordered_idx,
                                                         row.part_ordered_idx)
                                                    ], axis=1)
users

## Получение правильного упорядочивания элементов последовательностей по дате

%%time
users['reindex'] = users.date_sorting_val.apply(lambda x: list(np.argsort(x)))
users.head(10)

## Сохранение последовательностей в виде текстовых файлов

cols = ['url_host_idx', 'date_sorting_val', 'region_name_idx', 'city_name_idx',
       'cpe_manufacturer_name_idx', 'cpe_model_name_idx',
       'cpe_type_cd_idx', 'cpe_model_os_type_idx', 'price_idx', 'request_cnt']

for col in tqdm(cols):
    pbar = tqdm(total=users.shape[0])
    vals = users[col].apply(fillna)

    max_val = vals.apply(max).max() + 1

    with gzip.open('files/seqs_%s.gz'%col.split('_idx')[0], 'wt') as f:
        for i, reindex in enumerate(tqdm(users['reindex'].values)):
            row = np.array(vals[i])[reindex].copy()
            mask = row < 0
            row[mask] = max_val
            f.write(' '.join(map(str, row)))
            f.write('\n')

## Загрузка мешка слов для ссылок

with gzip.open('files/url_host_cbag_v2.pickle.gz', 'rb') as f:
    datamap = pickle.load(f)
datamap

## Формирование tsv файла с таргетами

trg_df = pd.DataFrame(dict(user_id=datamap['uids'])
            ).merge(pd.read_parquet('public_train.pqt'),
                    how='left')
trg_df.to_csv('target.tsv.gz', sep='\t', index=False)
trg_df.sample(10)

## Создание dummy файла скоров для удобства в будущем

sample_df = pd.read_parquet('submit_2.pqt')
sample_df = \
sample_df.merge(pd.DataFrame(dict(user_id=datamap['uids'],
                                  idx=np.arange(len(datamap['uids'])))),
                how='left')

sample_df['is_male'] = np.random.random(size=sample_df.idx.values.size)
sample_df['age'] = np.random.randint(1, 7, size=sample_df.idx.values.size)
sample_df = \
sample_df.sort_values(by='user_id').\
          drop('idx', axis=1)

sample_df.\
    to_csv('baseline_submission_wo_index.csv',
           sep=',',
           index=False,)

sample_df.\
    to_csv('baseline_submission_w_index.csv',
           sep=',',
           index=True,)
