==> 3 url bigrams.ipynb <==
## Импорт библиотек

%%time
import gzip
import pickle
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
from tqdm.auto import tqdm
%matplotlib inline
from scipy.sparse import csr_matrix, vstack
from IPython.display import clear_output
from ipywidgets.widgets import interact, interact_manual
import os
import gc
from PIL import Image
from multiprocessing.pool import Pool
from functools import partial
import sentencepiece as spm
import json
import gensim
from gensim.test.utils import datapath
import logging
from sklearn.decomposition import TruncatedSVD
logging.basicConfig(format='%(asctime)s : %(levelname)s : %(message)s', level=logging.INFO)

## Вывод всех файлов с последовательностями

!ls files/*seq*

## Загрузка последовательностей ссылок

XX = []
c = 0
with gzip.open('files/seqs_url_host.gz', 'rt', encoding='utf-8') as f:
    for line in tqdm(f):
        XX.append(list(map(int, line.strip().split(' '))))
        c += 1
len(XX)

## Загрузка последовательностей дат

TT = []
c = 0
with gzip.open('files/seqs_date_sorting_val.gz', 'rt', encoding='utf-8') as f:
    for line in tqdm(f):
        TT.append(list(map(int, line.strip().split(' '))))
        c += 1
len(TT)

## Статистики : mean-max-std кол-во дублей в последовательностях

stats = []
for tt in tqdm(TT):
    tmp = np.unique(tt, return_counts=True)[-1].copy()
    stats.append((tmp.max(), tmp.mean(), tmp.std()))
np.mean(stats, axis=0), np.max(stats, axis=0)

## Подсчёт частоты ссылок в последовательностях

idx_cnt_map = {}
for xx in tqdm(XX):
    for x in xx:
        idx_cnt_map[x] = idx_cnt_map.get(x, 0) + 1
len(idx_cnt_map)

## Оставляем элементы последовательностей с большой частотой

XX_ = []
TT_ = []
pbar = tqdm(total=len(XX))
for xx, tt in zip(XX, TT):
    keep_mask = [idx_cnt_map.get(xxx, 0) >= 50 for xxx in xx]
    XX_.append([xxx for xxx, remain in zip(xx, keep_mask) if remain])
    TT_.append([ttt for ttt, remain in zip(tt, keep_mask) if remain])
    pbar.update(1)
del XX, TT
gc.collect()
max(map(len, XX_))

## Статистики уменьшились

stats = []
for tt in tqdm(TT_):
    if not len(tt):
        stats.append((0, 0, 0))
        continue
    tmp = np.unique(tt, return_counts=True)[-1].copy()
    stats.append((tmp.max(), tmp.mean(), tmp.std()))
np.mean(stats, axis=0), np.max(stats, axis=0)

## Считаем частоты встречаемости пар ссылок в наших данных

interaction_cnt_map = dict()
pbar = tqdm(total=len(XX_))
for xx, tt in zip(XX_, TT_):
    if not len(xx):
        #II.append([])
        pbar.update(1)
        continue
    tmp = np.array(tt)
    stoppers = list(np.where(tmp[:-1] != tmp[1:])[0] + 1)
    stoppers = [0] + stoppers + [len(tt)]
    xx = np.array(xx)
    segments = []

    for left, right in zip(stoppers[:-1], stoppers[1:]):
        segments.append(set(list(xx[left:right])))

    for left_segment, right_segment in zip(segments[:-1], segments[1:]):
        for i in left_segment:
            for j in right_segment:
                # тут сортируем индексы ссылок, чтобы уменьшить разнообразие
                key = tuple(sorted([i, j]))
                interaction_cnt_map[key] = interaction_cnt_map.get(key, 0) + 1
    pbar.update(1)

## Сколько всего получилось пар ссылок

len(interaction_cnt_map)

## Сохраняем промежуточный результат на всякий случай

%%time
with gzip.open('url_bigrams_keys.pickle.gz', 'wb') as f:
    pickle.dump(interaction_cnt_map, f, protocol=-1)

## Оставляем только частотные пары ссылок

interaction_whitelist = set([k for k, v in interaction_cnt_map.items() if v>=80])
interaction_whitelist = {k: i for i, k in enumerate(interaction_whitelist)}
len(interaction_whitelist)

## Считаем биграммы на декартовых произведениях соседних временных множест

output_folder = 'auxilary/bigrams'
os.makedirs(output_folder, exist_ok=True)
II_rows = []
II_cols = []
II_data = []
row_offset = 0
CHUNK_SIZE = 50_000
pbar = tqdm(total=len(XX_))
for row_id, (xx, tt) in enumerate(zip(XX_, TT_)):
    if not len(xx):
        pbar.update(1)
        continue
    tmp = np.array(tt)
    stoppers = list(np.where(tmp[:-1] != tmp[1:])[0] + 1)
    stoppers = [0] + stoppers + [len(tt)]
    xx = np.array(xx)
    segments = []

    for left, right in zip(stoppers[:-1], stoppers[1:]):
        segments.append(set(list(xx[left:right])))

    ii_cols = []
    for left_segment, right_segment in zip(segments[:-1], segments[1:]):
        for i in left_segment:
            for j in right_segment:
                key = tuple(sorted([i, j]))
                if key not in interaction_whitelist: continue
                ii_cols.append(interaction_whitelist[key])
    II_cols.extend(ii_cols)
    II_rows.extend([row_id - row_offset]*len(ii_cols))
    II_data.extend([1]*len(ii_cols))
    pbar.update(1)
    if (row_id % CHUNK_SIZE == 0) and row_id:
        tmp_dat = csr_matrix((II_data, (II_rows, II_cols)),
                             shape=(II_rows[-1] + 1,
                                    len(interaction_whitelist)),
                             dtype=np.uint32)
        with gzip.open(os.path.join(output_folder, '%06d.pickle.gz'%row_id), 'wb') as f:
            pickle.dump(tmp_dat, f, protocol=-1)
        row_offset = row_id + 1
        II_rows = []
        II_cols = []
        II_data = []
if len(II_rows):
    tmp_dat = csr_matrix((II_data, (II_rows, II_cols)),
                         shape=(II_rows[-1] + 1,
                                len(interaction_whitelist)),
                         dtype=np.uint16)
    with gzip.open(os.path.join(output_folder, '%d.pickle.gz'%row_id), 'wb') as f:
        pickle.dump(tmp_dat, f, protocol=-1)
len(II_data)

## Склеиваем чанки в одну sparse матрицу

%%time
sparse_dats = []
for file in tqdm(sorted(os.listdir(output_folder))):
    if 'ipynb_checkpoints' in file:
        continue
    print(file)
    with gzip.open(os.path.join(output_folder, file), 'rb') as f:
        sparse_dats.append(pickle.load(f))
sparse_dats[2] = \
csr_matrix((sparse_dats[2].data,
            sparse_dats[2][1:].nonzero()),
           shape=(50000, sparse_dats[2].shape[1]))
sparse_dats[-1] = \
csr_matrix((sparse_dats[-1].data,
            sparse_dats[-1][1:].nonzero()),
           shape=(sparse_dats[-1].shape[0], sparse_dats[-1].shape[1]))

sparse_data = vstack(sparse_dats[:1] +
                     [(sd[1:] if sd[0].data.size==0 else sd) for sd in sparse_dats[1:]])

sparse_data.shape

## Проверяем на случайном поднаборе последовательностей, что биграмы упорядочены правильно

checks = 0
for row_id, (xx, tt) in tqdm(enumerate(zip(XX_, TT_))):
    if not len(xx):
        print(row_id)
        continue
    if np.random.random() < 0.98: continue
    tmp = np.array(tt)
    stoppers = list(np.where(tmp[:-1] != tmp[1:])[0] + 1)
    stoppers = [0] + stoppers + [len(tt)]
    xx = np.array(xx)
    segments = []

    for left, right in zip(stoppers[:-1], stoppers[1:]):
        segments.append(set(list(xx[left:right])))

    ii_cols = []
    for left_segment, right_segment in zip(segments[:-1], segments[1:]):
        for i in left_segment:
            for j in right_segment:
                key = tuple(sorted([i, j]))
                if key not in interaction_whitelist: continue
                ii_cols.append(interaction_whitelist[key])
    assert set(ii_cols) == set(list(sparse_data[row_id].nonzero()[-1]))
    checks += 1
checks

## Сохраняем биграмы как разряженную матрицу

%%time
with gzip.open('auxilary/bigrams_sparse.pickle.gz', 'wb') as f:
    pickle.dump(sparse_data, f, protocol=-1)

## До этого фильтровали по кол-ву элементов, а сейчас по числу уникальных юзеров

feats_mask = (np.array((sparse_data>0).sum(axis=0)).flatten() > 40)
feats_mask.sum()

## Понижаем размерность с 580к до 512 признаков

%%time
sub_data = sparse_data[:, feats_mask]
sub_data = csr_matrix(
    (sub_data.data**0.5, sub_data.nonzero()), shape=sub_data.shape, dtype=np.float32
)
latents = \
TruncatedSVD(n_components=512, random_state=42
            ).fit_transform(sub_data)

## Сохраняем

%%time
with gzip.open('auxilary/bigrams_dense.pickle.gz', 'wb') as f:
    pickle.dump(latents, f, protocol=-1)
