==> 4_1 data prepare.ipynb <==
## Установка CLIP

!pip install git+https://github.com/openai/CLIP.git

## Импорт библиотек

import os
import json
from tqdm.auto import tqdm
import time
import gzip
from PIL import Image
import numpy as np
import pickle
from IPython.display import display_png

import torch
import clip

## Определение директорий с артефактами парсинга

FOLDERS = ['desktop_chunks20k']

## Загрузка модели CLIP

device = "cuda" if torch.cuda.is_available() else "cpu"
model, preprocess = clip.load("ViT-L/14", device=device, download_root='CLIP')
model

## Получение эмбеддингов CLIP

processed_urls = []
preds = []
total_pbar = tqdm()
batch = []
savefile = 'clipVIT_scores_20k.pickle.gz'
for folder in FOLDERS:
    for cnt, file in tqdm(enumerate(map(lambda x: os.path.join(folder, x), os.listdir(folder)))):
        try:
            with gzip.open(file, 'rb') as f:
                chunk = pickle.load(f)
        except:
            continue
        if len(chunk['labels']):
            processed_urls.extend(list(chunk['labels'][:, 0]))
            for img in chunk['data']:
                with torch.no_grad():
                    preds.append(
                        model.encode_image(
                            preprocess(
                                Image.fromarray(img).convert('RGB')
                            ).unsqueeze(0).to(device)
                        ).detach().cpu().numpy().flatten()
                    )
                total_pbar.update(1)
        if cnt % 1000 == 0:
            with gzip.open(savefile, 'wb') as f:
                pickle.dump(dict(domains=processed_urls, scores=preds), f, protocol=-1)
with gzip.open(savefile, 'wb') as f:
    pickle.dump(dict(domains=processed_urls, scores=preds), f, protocol=-1)

## Получение тайтлов ссылок

processed_urls = []
total_pbar = tqdm()
for folder in FOLDERS:
    for cnt, file in tqdm(enumerate(map(lambda x: os.path.join(folder, x), os.listdir(folder)))):
        try:
            with gzip.open(file, 'rb') as f:
                chunk = pickle.load(f)
        except:
            continue
        if len(chunk['labels']):
            for (domain, title), html in zip(chunk['labels'], chunk['html']):
                processed_urls.append((domain, title, html))
                total_pbar.update(1)
with gzip.open('domain20k_text.pickle.gz', 'wb') as f:
    pickle.dump(dict(domain=np.array([x[0] for x in processed_urls]),
                     title=np.array([x[1] for x in processed_urls])), f, protocol=-1)

## Получение HTML ссылок

with gzip.open('domain20k_html.txt.gz', 'wt', encoding='utf-8') as f:
    for x in tqdm(processed_urls):
        if '\n=\n=\n' in x[2]:
            print('!!!', x[0], x[1])
            break
        f.write(x[2]+'\n=\n=\n')
