==> 4_2 simweb.ipynb <==
## Импорт библиотек

import pandas as pd
import numpy as np
import gzip
import os
import sys
from bs4 import BeautifulSoup
from tqdm.auto import tqdm
from multiprocessing.pool import Pool
import re
import pickle
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.feature_extraction.text import TfidfTransformer
from scipy.sparse import csr_matrix
from sklearn.preprocessing import Normalizer

## Загрузка файлов

def worker(file):
    try:
        with gzip.open('simwebhtml/%s'%file, 'rb') as f:
            page_source = f.read().decode('utf-8')
        soup = BeautifulSoup(page_source, 'html.parser')
        info = soup.find(id='demographics').text
        domain = file[:-3]
        good_file = 'wa-demographics__age-data-label' in page_source
        return (domain, soup.find_all('meta')[3].get('content', ''), info, good_file)
    except Exception as e:
        e = sys.exc_info()
        if isinstance(e[0], KeyboardInterrupt):
            return None
        return None


info = []
with Pool(8) as pool:
    files = os.listdir('simwebhtml')
    pbar = tqdm(total = len(files))
    for result in pool.imap(worker, files):
        if result is None: continue
        info.append(
            result
        )
        pbar.update(1)

## Получение категории и рейтинга сайта

regex = re.compile(
    r'^([\w\d-]*\.[\w\d-]*) is ranked #(\d+) in the ([\w\s\>-]+) category and #(\d+) Globally.*$'
)

meta_info = []
for x in tqdm(info):
    res = regex.findall(x[1].replace('&', 'and'))[0]
    assert len(res) == 4
    (domain, top_cat, cat, top_all) = res
    meta_info.append((domain, int(top_cat), cat.strip(), int(top_all)))

meta_info

## Пример информации о сайте

info[2]

## Выделение информации о поле

ismale = []
err_cnts = 0
for x in tqdm(info):
    try:
        ismale.append(float(x[2].split('audience is ')[-1].split('%')[0]))
    except Exception as e:
        err_cnts += 1
        ismale.append(50.)
err_cnts

## Выделение информации о возрасте, мода

age_group = []
err_cnts = 0
for x in tqdm(info):
    try:
        age = str(x[2].split(
            'The largest age group of visitors are '
        )[-1].split(' year olds')[0]).replace(' ', '')
        assert len(age) < 10
        age_group.append(age)
    except Exception as e:
        err_cnts += 1
        age_group.append('25-34')
err_cnts

## Выделение информации о возрасте, гистограма

age_probs = []
err_cnts = 0
for x in tqdm(info):
    try:
        probs =  list(map(float, 
            x[2].split('Highcharts 10.3.2')[-1].split('18 - 24')[0].split('%')[:-1]
                         ))
        assert len(probs) == 6
        age_probs.append(
           probs
        )
    except Exception as e:
        err_cnts += 1
        age_probs.append([0]*6)
age_probs = np.array(age_probs)
age_probs[age_probs[:, 0]==0] = age_probs[np.unique(age_probs.nonzero()[0])].mean(axis=0)
err_cnts

## Векторизация категорий

topics = [x[2] for x in meta_info]

topics_cbag = CountVectorizer(min_df=10).fit_transform(pd.Series(topics))
topics_cbag

## Векторизация возраста и пола

with gzip.open('files/url_host_cbag_v2.pickle.gz', 'rb') as f:
    datamap = pickle.load(f)
datamap

map_df = pd.read_csv('auxilary/url_host_mapper_v2.tsv.gz', sep='\t')

reindex = \
pd.DataFrame(
    dict(url_host=[x[0] for x in info])
        ).merge(map_df).drop_duplicates('url_host').url_host_idx.values.copy()

all_data = datamap['data'][:, reindex]
all_data_sqrt = csr_matrix((all_data.data**0.25, all_data.nonzero()),
                     shape=all_data.shape,
                     dtype=np.float32)

sex_score = TfidfTransformer(sublinear_tf=True).fit_transform(
        all_data_sqrt
).dot(np.array(ismale)-50.)
sex_score = sex_score/sex_score.std()/3.

topics_score = TfidfTransformer(sublinear_tf=True).fit_transform(
        all_data_sqrt
).dot(topics_cbag.toarray() - topics_cbag.toarray().mean(axis=0))
topics_score = Normalizer().fit_transform(topics_score)
topics_score = topics_score / topics_score.std(axis=0)
topics_score = np.clip(topics_score, -5, 5)
topics_score

top_rank = \
    np.array([x[1] for x in meta_info]).argsort().argsort()/len(meta_info) - 0.5
ranking_score = TfidfTransformer(sublinear_tf=True).fit_transform(
        all_data_sqrt
).dot(np.array(top_rank))
ranking_score = ranking_score/ranking_score.std()/3.

all_rank = \
    np.array([x[3] for x in meta_info]).argsort().argsort()/len(meta_info) - 0.5
ranking_score2 = TfidfTransformer(sublinear_tf=True).fit_transform(
        all_data_sqrt
).dot(np.array(all_rank))
ranking_score2 = ranking_score2/ranking_score2.std()/3.

age_score = \
TfidfTransformer(sublinear_tf=True).fit_transform(
        all_data_sqrt
).dot(pd.get_dummies(age_group).values[:, :-1])
age_score = Normalizer().fit_transform(age_score)

age_score2 = \
TfidfTransformer(sublinear_tf=True).fit_transform(
        all_data_sqrt
).dot(age_probs - age_probs.mean(axis=0))
age_score2 = Normalizer().fit_transform(age_score2)
age_score2

## Конкатенация признаков SimWeb

simweb_scores = \
np.concatenate([sex_score[:, None],
                topics_score,
                ranking_score[:, None],
                ranking_score2[:, None],
                age_score,
                age_score2
                ], axis=1)

simweb_scores

## Сохранение признаков

with gzip.open('auxilary/simweb_domain.pickle.gz', 'wb') as f:
    pickle.dump(simweb_scores, f, protocol=-1)
