==> 8 nn stacking.ipynb <==
## Ограничения на ресурсы для numpy, импорт библиотек

import os
os.environ["OMP_NUM_THREADS"] = "4" # export OMP_NUM_THREADS=4
os.environ["OPENBLAS_NUM_THREADS"] = "4" # export OPENBLAS_NUM_THREADS=4 
os.environ["MKL_NUM_THREADS"] = "6" # export MKL_NUM_THREADS=6
os.environ["VECLIB_MAXIMUM_THREADS"] = "4" # export VECLIB_MAXIMUM_THREADS=4
os.environ["NUMEXPR_NUM_THREADS"] = "6" # export NUMEXPR_NUM_THREADS=6

import pandas as pd
import numpy as np
import os
import time
from tqdm.auto import tqdm
from functools import reduce

import scipy
from scipy.sparse import csr_matrix, hstack, vstack
from scipy.optimize import minimize
import gzip
import pickle

import sklearn
from sklearn.utils import resample
from sklearn.pipeline import Pipeline
from sklearn.preprocessing import KBinsDiscretizer
from sklearn.preprocessing import QuantileTransformer, Normalizer
from sklearn.feature_extraction.text import TfidfTransformer
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.linear_model import LogisticRegression
from sklearn.model_selection import cross_val_predict, StratifiedKFold
from sklearn.ensemble import BaggingClassifier
from sklearn.metrics import roc_auc_score, classification_report, confusion_matrix, f1_score
from sklearn.linear_model import LogisticRegression, Ridge
from sklearn.model_selection import cross_val_predict

import matplotlib.pyplot as plt
%matplotlib inline

## Загрузка мешка слов по ссылкам

with gzip.open('files/url_host_cbag_v2.pickle.gz', 'rb') as f:
    datamap = pickle.load(f)
datamap

## Получение частотных ссылок мешка слов url_host

feats_mask = (np.array((datamap['data']>0).sum(axis=0)).flatten() > 40)
feats_mask.sum()

## Считывание файла с таргетами

trg_df = pd.read_csv('target.tsv.gz', sep='\t')
trg_df.sample(10)

# Подготовка таргетов и поднабора юзеров из обучения

key = 'data'

(trg_df.age.isna()|trg_df.is_male.isna()).sum(),\

all_mask = (~trg_df.age.isna()|~trg_df.is_male.isna()).values.copy()
trg_train = trg_df[all_mask].fillna({'is_male': 0.5, 'age':34})
trg_age = trg_train.age.values.copy()
trg_sex = trg_train.is_male.values.copy()


age_bins = [[0, 25], [26, 35], [36, 45], [46, 55], [56, 65], [66, 999]]

print('Train sample:', all_mask.sum())

y_all = 0
for k, age_bin in enumerate(age_bins):
    y = pd.Series(trg_age).between(*age_bin).values.copy()
    y_all += y*(k+1)

X_tr = datamap[key][all_mask][:, feats_mask]

y_all.min(), y_all.max(), X_tr.shape

# Настройка валидации

RANDOM_SEED = 42
kfold = StratifiedKFold(n_splits=10,
                        shuffle=True,
                        random_state=RANDOM_SEED)
folds = [(train_ind, test_ind) for train_ind, test_ind in
         kfold.split((np.uint8(trg_sex*2)+y_all*10).astype(str),
                     (np.uint8(trg_sex*2)+y_all*10).astype(str))]

# Загрузка скоров

all_preds_sex = []
all_preds_age = []
model_names = []
for k_start, k_end, file in [(1, 4, 'preds_290323_age_p'),
                             (1, 6, 'preds_290323_sex_p'),
                             (1, 5, 'preds_280323_sex_p'),
                             (1, 6, 'preds_280323_age_p'),
                             (1, 7, 'preds_270323_sex_p'),
                             (1, 10, 'preds_270323_age_p'),
                             (1, 21, 'preds_230323_age_p'),
                             (1, 16, 'preds_230323_sex_p'),
                             (1, 18, 'preds_120323_sex_p'),
                             (1, 14, 'preds_120323_age_p'),
                             (1, 14, 'preds_100323_sex_p'),
                             (1, 12, 'preds_100323_age_p'),
                             (1, 9, 'preds_070323_age_p'),
                             (1, 9, 'preds_070323_sex_p'),
                            ]:
    tmp_preds_sex = []
    tmp_preds_age = []
    for k in tqdm(range(k_start, k_end)):
        filename = '%s%d.pickle.gz'%(file, k)
        with gzip.open(filename, 'rb') as f:
            predsmap = pickle.load(f)
            tmp_preds_age.append(predsmap['age'])
            tmp_preds_sex.append(predsmap['sex'])
            model_names.append(filename)
    all_preds_sex.extend(tmp_preds_sex)
    all_preds_age.extend(tmp_preds_age)


## Лучшая модель по полу

roc_aucs = []
for probs in tqdm(all_preds_sex):
    roc_aucs.append(
        roc_auc_score(trg_sex[trg_sex!=0.5],
                      probs[all_mask][trg_sex!=0.5])
    )
model_names[np.argmax(roc_aucs)], np.max(roc_aucs)

## Лучшая модель по возрасту

f1s = []
for probs in tqdm(all_preds_age):
    f1s.append(
        f1_score(y_all-1,
             probs.argmax(axis=1)[all_mask],
             average='weighted'
        )
    )
model_names[np.argmax(f1s)], np.max(f1s)

## Лучший скор 1 модели

np.max(f1s)*2 + np.max(roc_aucs)*2 - 1

## Stacking Logreg для пола 

lr = \
LogisticRegression(C=1e-2, penalty='l2',
                   solver='liblinear', max_iter=10)

probs = \
cross_val_predict(lr,
                  np.array(all_preds_sex
                                 ).T.reshape(-1,
                                             len(all_preds_sex)
                                            )[all_mask],
                  trg_sex.astype(int),
                  cv=folds,
                  verbose=2,
                  n_jobs=len(folds),
                  method='predict_proba')[:, 1]

lr.fit(np.array(all_preds_sex
             ).T.reshape(-1,
                         len(all_preds_sex)
                        )[all_mask],
       trg_sex.astype(int))

cv_scores = []
for _, test_ind in folds:
    probs[test_ind] = probs[test_ind].argsort().argsort() / test_ind.size
    cv_score = \
    roc_auc_score(trg_sex[test_ind][trg_sex[test_ind]!=0.5],
                        probs[test_ind][trg_sex[test_ind]!=0.5])
    cv_scores.append(cv_score)
    print(cv_score)

print('='*50)

print(roc_auc_score(trg_sex[trg_sex!=0.5], probs[trg_sex!=0.5]))

## Взвешивание скоров пола по коэффициентам логрега

all_preds_sex = lr.coef_.flatten().dot(all_preds_sex)

## Stacking One-VS-Rest Ridge для возраста

scipy_tmp = np.array(all_preds_age).T
print(scipy_tmp.shape)
age_coefs = []
age_scores = []

scipy_tmp_all = np.array(scipy_tmp).transpose([1, 0, 2]).reshape(all_preds_sex.size, -1)
age_probs_cv = np.zeros((all_mask.sum(), len(age_bins)))
age_probs_train = np.zeros((all_mask.sum(), len(age_bins)))
age_probs_all = np.zeros((all_preds_sex.size, len(age_bins)))

for trg_val in tqdm(range(1, 7)):
    trg_pair = [trg_val - 1]
    age_X_tr = np.concatenate(scipy_tmp[trg_pair].transpose([0, 2, 1]), axis=0).T
    for train_ind, test_ind in tqdm(folds):
        lr = Ridge(alpha=15,
                   fit_intercept=True)

        lr.fit(age_X_tr[all_mask][train_ind],
               (y_all==trg_val)[train_ind])

        tmp_preds = lr.predict(age_X_tr) - lr.intercept_
        age_probs_cv[test_ind, trg_val-1] = tmp_preds[all_mask][test_ind]
        age_probs_train[train_ind, trg_val-1] = tmp_preds[all_mask][train_ind]
        age_probs_all[:, trg_val-1] += tmp_preds

cm = confusion_matrix(y_all - 1, age_probs_cv.argmax(axis=1))
# распределение бинов возраста в public_lb
true_weights = np.array([11.970, 32.675, 28.855, 15.886, 8.643, 1.970]) / 1e2
print((2 * np.diag(cm) / (cm.sum(axis=1) + cm.sum(axis=0))).dot(true_weights))

## Оценка распределения бинов на private_lb

# восстанавливаем границы бинов как они указаны в правилах
age_bins_ = age_bins.copy()
age_bins_[0] = [19, 25]
age_labels, age_cnts = \
np.unique(
    np.concatenate([trg_df.age.between(*age_bin).values[:, None] for age_bin in age_bins_],
               axis=1).argmax(axis=1)[all_mask],
    return_counts=True
)
# оценка распределения бинов возраста как взвешенная сумма распределений на train и public
# в train 270к примеров, по оценке в паблик 80к примеров
age_cnts_estimator = \
age_cnts / age_cnts.sum() * 270_000 + true_weights * 80_000

true_weights, \
age_cnts / age_cnts.sum(), \
age_cnts_estimator / age_cnts_estimator.sum()

## Финальный подбор вектора-смещений, максимизирующий f1_w

def maximize_public_f1(bias):
    bias = np.round(bias, decimals=4)
    cm = confusion_matrix(y_all - 1,
                          (age_probs_cv + bias).argmax(axis=1))
    true_weights = age_cnts_estimator / age_cnts_estimator.sum()
    return -(2 * np.diag(cm) / (cm.sum(axis=1) + cm.sum(axis=0))).dot(true_weights)

scipy_eps = 1e-10
scipy_res = None
bf1 = 0
bias = 0.
max_iters = 2000
# стохастический поиск для инициализации смещения
for n_iter in tqdm(range(max_iters)):
    x = np.random.normal(size=len(age_bins)) \
        / np.random.choice([50., 100., 150., 200.])
    x += bias * (1 + x)
    fun = maximize_public_f1(x)
    if abs(fun) > bf1:
        print(abs(fun), x)
        bf1 = abs(fun)
        bias = np.round(x, decimals=4)
# уточнения смещения с помощью COBYLA
for _ in range(10):
    scipy_res = minimize(maximize_public_f1,
                     bias,
                     method='COBYLA')
    print(abs(scipy_res.fun), scipy_res.x)
    if abs(scipy_res.fun) > bf1:
        bf1 = abs(scipy_res.fun)
        bias = np.round(scipy_res.x, decimals=4)

## Получение бинов возраста

all_preds_age_label = (age_probs_all + bias).argmax(axis=1) + 1
all_preds_age_label

## Скор после stacking'а

2 * bf1 + 2 * roc_auc_score(trg_sex[trg_sex!=0.5], probs[trg_sex!=0.5]) - 1

## Формирование посылки

sample_df = pd.read_csv('baseline_submission_wo_index.csv')
sample_df = \
sample_df.merge(pd.DataFrame(dict(user_id=datamap['uids'],
                                  idx=np.arange(len(datamap['uids'])))),
                how='left')

sample_df['is_male'] = all_preds_sex[sample_df.idx.values].argsort().argsort()/sample_df.shape[0]
sample_df['age'] = all_preds_age_label[sample_df.idx.values]
sample_df = \
sample_df.sort_values(by='user_id').\
          drop('idx', axis=1)

sample_df.\
    to_csv('20230329r_submission_wo_index.csv',
           sep=',',
           index=False)
